additional functionality extraction for pof
