+++ src/main/java/menagerie/gui/itemhandler/opener/ItemOpener.java
@@ -1,0 +1,8 @@
+package menagerie.gui.itemhandler.opener;
+
+import menagerie.gui.MainController;
+import menagerie.model.menagerie.Item;
+
+public interface ItemOpener {
+  void open(Item item, MainController controller);
+}
